requirements.txt
pyaudio
python-dotenv
sense-hat
openai
requests